+++ qbraid/runtime/ibm/provider.py
@@ -31,7 +31,7 @@ from .device import QiskitBackend
 if TYPE_CHECKING:
     import qiskit_ibm_runtime
 
-    import qbraid.runtime.qiskit
+    import qbraid.runtime.ibm
 
 
 class QiskitRuntimeProvider(QuantumProvider):
@@ -100,7 +100,7 @@ class QiskitRuntimeProvider(QuantumProvider):
         )
 
     @cached_method
-    def get_devices(self, operational=True, **kwargs) -> list[qbraid.runtime.qiskit.QiskitBackend]:
+    def get_devices(self, operational=True, **kwargs) -> list[qbraid.runtime.ibm.QiskitBackend]:
         """Returns the IBM Quantum provider backends."""
 
         backends = self.runtime_service.backends(operational=operational, **kwargs)
@@ -116,7 +116,7 @@ class QiskitRuntimeProvider(QuantumProvider):
     @cached_method
     def get_device(
         self, device_id: str, instance: Optional[str] = None
-    ) -> qbraid.runtime.qiskit.QiskitBackend:
+    ) -> qbraid.runtime.ibm.QiskitBackend:
         """Returns the IBM Quantum provider backends."""
         backend = self.runtime_service.backend(device_id, instance=instance)
         return QiskitBackend(
@@ -125,7 +125,7 @@ class QiskitRuntimeProvider(QuantumProvider):
 
     def least_busy(
         self, simulator=False, operational=True, **kwargs
-    ) -> qbraid.runtime.qiskit.QiskitBackend:
+    ) -> qbraid.runtime.ibm.QiskitBackend:
         """Return the least busy IBMQ QPU."""
         backend = self.runtime_service.least_busy(
             simulator=simulator, operational=operational, **kwargs