+++ tests/top_level/test_caching.py
@@ -12,21 +12,33 @@
 Unit tests for caching module.
 
 """
+import math
+
 import pytest
 
-from qbraid._caching import _generate_cache_key, cached_method
+from qbraid._caching import _generate_cache_key, cached_method, clear_cache
 
 
 class TestClass:
+    """Test class for cached_method decorator."""
+
+    def __init__(self, x: int, y: int):
+        self.x = x
+        self.y = y
 
     @cached_method
-    def get_data(self, param: int) -> int:
-        return param * 2
+    def adjusted_factorial(self, n: int) -> int:
+        """Calculate the factorial of n, adjusted by the instance attrs x and y."""
+        base_factorial = math.factorial(n + self.x)
+        return base_factorial**self.y
+
+    def __hash__(self):
+        return hash((self.x, self.y))
 
 
 @pytest.fixture
 def test_instance():
-    return TestClass()
+    return TestClass(1, 2)
 
 
 def test_generate_cache_key(test_instance):
@@ -34,3 +46,19 @@ def test_generate_cache_key(test_instance):
     key = _generate_cache_key(test_instance, "get_data", (1,), {})
     assert isinstance(key, str)
     assert len(key) == 64  # SHA-256 hash length
+
+
+def test_clear_cache(test_instance, monkeypatch):
+    """Test clearing the LRU cache."""
+    monkeypatch.setenv("DISABLE_CACHE", "0")
+
+    assert test_instance.adjusted_factorial.cache_info().currsize == 0
+
+    for i in range(3):
+        test_instance.adjusted_factorial(i)
+
+    assert test_instance.adjusted_factorial.cache_info().currsize == 3
+
+    clear_cache()
+
+    assert test_instance.adjusted_factorial.cache_info().currsize == 0